优化笔记编辑器

+++ src/components/TrashDetail.vue
@@ -53,6 +53,10 @@ export default {
     this.getTrashNotes()
       .then(() => {
         this.setCurTrashNote({curTrashNoteId: this.$route.query.noteId});
+        this.$router.replace({
+          path: '/trash',
+          query: {noteId: this.curTrashNote.id}
+        });
       });
   },
 
@@ -79,11 +83,30 @@ export default {
       'getNotebooks'
     ]),
     onDelete() {
-      console.log({noteId: this.curTrashNote.id});
-      this.deleteTrashNote({noteId: this.curTrashNote.id});
+      this.$confirm('删除后将无法恢复', '确定删除？', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        return this.deleteTrashNote({noteId: this.curTrashNote.id});
+      }).then(() => {
+        console.log('delete success');
+        this.setCurTrashNote();
+        this.$router.replace({
+          path: '/trash',
+          query: {noteId: this.curTrashNote.id}
+        });
+      });
     },
     onRevert() {
-      this.revertTrashNote({noteId: this.curTrashNote.id});
+      this.revertTrashNote({noteId: this.curTrashNote.id})
+        .then(() => {
+          this.setCurTrashNote();
+          this.$router.replace({
+            path: '/trash',
+            query: {noteId: this.curTrashNote.id}
+          });
+        });
     }
   },
   beforeRouteUpdate(to, from, next) {
